+++ example/lib/models/post.g.dart
@@ -124,9 +124,15 @@ AutoDisposeStateNotifierProvider<DataStateNotifier<List<Post>>,
     watchPosts(
         {bool? remote,
         Map<String, dynamic>? params,
-        Map<String, String>? headers}) {
-  return _watchPosts(
-      WatchArgs(remote: remote, params: params, headers: headers));
+        Map<String, String>? headers,
+        bool Function(Post)? filterLocal,
+        bool? syncLocal}) {
+  return _watchPosts(WatchArgs(
+      remote: remote,
+      params: params,
+      headers: headers,
+      filterLocal: filterLocal,
+      syncLocal: syncLocal));
 }
 
 extension PostX on Post {
@@ -134,8 +140,10 @@ extension PostX on Post {
   /// [save], [delete] and so on.
   ///
   /// Can be obtained via `context.read`, `ref.read`, `container.read`
-  Post init(Reader read) {
+  Post init(Reader read, {bool save = true}) {
     final repository = internalLocatorFn(postsRepositoryProvider, read);
-    return repository.remoteAdapter.initializeModel(this, save: true);
+    final updatedModel =
+        repository.remoteAdapter.initializeModel(this, save: save);
+    return save ? updatedModel : this;
   }
 }