+++ example/lib/models/user.g.dart
@@ -102,9 +102,15 @@ AutoDisposeStateNotifierProvider<DataStateNotifier<List<User>>,
     watchUsers(
         {bool? remote,
         Map<String, dynamic>? params,
-        Map<String, String>? headers}) {
-  return _watchUsers(
-      WatchArgs(remote: remote, params: params, headers: headers));
+        Map<String, String>? headers,
+        bool Function(User)? filterLocal,
+        bool? syncLocal}) {
+  return _watchUsers(WatchArgs(
+      remote: remote,
+      params: params,
+      headers: headers,
+      filterLocal: filterLocal,
+      syncLocal: syncLocal));
 }
 
 extension UserX on User {
@@ -112,8 +118,10 @@ extension UserX on User {
   /// [save], [delete] and so on.
   ///
   /// Can be obtained via `context.read`, `ref.read`, `container.read`
-  User init(Reader read) {
+  User init(Reader read, {bool save = true}) {
     final repository = internalLocatorFn(usersRepositoryProvider, read);
-    return repository.remoteAdapter.initializeModel(this, save: true);
+    final updatedModel =
+        repository.remoteAdapter.initializeModel(this, save: save);
+    return save ? updatedModel : this;
   }
 }