+++ lib/src/repository/local_adapter.dart
@@ -16,13 +16,9 @@ abstract class LocalAdapter<T extends DataModel<T>> with _Lifecycle {
   // protected API
 
   /// Returns all models of type [T] in local storage.
-  @protected
-  @visibleForTesting
   List<T> findAll();
 
   /// Finds model of type [T] by [key] in local storage.
-  @protected
-  @visibleForTesting
   T? findOne(String key);
 
   /// Saves model of type [T] with [key] in local storage.