+++ test/_support/book.dart
@@ -5,30 +5,33 @@ part 'book.freezed.dart';
 part 'book.g.dart';
 
 @freezed
-@DataRepository([AuthorAdapter], remote: false)
-class Author with DataModel<Author>, _$Author {
-  @With.fromString('DataModel<Author>')
-  factory Author({
+@DataRepository([BookAuthorAdapter], remote: false)
+class BookAuthor with DataModel<BookAuthor>, _$BookAuthor {
+  @With.fromString('DataModel<BookAuthor>')
+  factory BookAuthor({
     required int id,
     String? name,
     HasMany<Book>? books,
-  }) = _Author;
-  factory Author.fromJson(Map<String, dynamic> json) => _$AuthorFromJson(json);
+  }) = _BookAuthor;
+  factory BookAuthor.fromJson(Map<String, dynamic> json) =>
+      _$BookAuthorFromJson(json);
 }
 
 @freezed
 @DataRepository([])
 class Book with DataModel<Book>, _$Book {
   @With.fromString('DataModel<Book>')
+  @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
   factory Book({
     required int id,
     String? title,
-    BelongsTo<Author>? author,
+    @Default(0) int numberOfSales,
+    @JsonKey(name: 'original_author') BelongsTo<BookAuthor>? originalAuthor,
   }) = _Book;
   factory Book.fromJson(Map<String, dynamic> json) => _$BookFromJson(json);
 }
 
-mixin AuthorAdapter on RemoteAdapter<Author> {
+mixin BookAuthorAdapter on RemoteAdapter<BookAuthor> {
   @override
   String get type => 'writers';
 }