+++ test/_support/book.freezed.dart
@@ -12,55 +12,57 @@ T _$identity<T>(T value) => value;
 final _privateConstructorUsedError = UnsupportedError(
     'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
-Author _$AuthorFromJson(Map<String, dynamic> json) {
-  return _Author.fromJson(json);
+BookAuthor _$BookAuthorFromJson(Map<String, dynamic> json) {
+  return _BookAuthor.fromJson(json);
 }
 
 /// @nodoc
-class _$AuthorTearOff {
-  const _$AuthorTearOff();
+class _$BookAuthorTearOff {
+  const _$BookAuthorTearOff();
 
-  _Author call({required int id, String? name, HasMany<Book>? books}) {
-    return _Author(
+  _BookAuthor call({required int id, String? name, HasMany<Book>? books}) {
+    return _BookAuthor(
       id: id,
       name: name,
       books: books,
     );
   }
 
-  Author fromJson(Map<String, Object> json) {
-    return Author.fromJson(json);
+  BookAuthor fromJson(Map<String, Object> json) {
+    return BookAuthor.fromJson(json);
   }
 }
 
 /// @nodoc
-const $Author = _$AuthorTearOff();
+const $BookAuthor = _$BookAuthorTearOff();
 
 /// @nodoc
-mixin _$Author {
+mixin _$BookAuthor {
   int get id => throw _privateConstructorUsedError;
   String? get name => throw _privateConstructorUsedError;
   HasMany<Book>? get books => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
-  $AuthorCopyWith<Author> get copyWith => throw _privateConstructorUsedError;
+  $BookAuthorCopyWith<BookAuthor> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $AuthorCopyWith<$Res> {
-  factory $AuthorCopyWith(Author value, $Res Function(Author) then) =
-      _$AuthorCopyWithImpl<$Res>;
+abstract class $BookAuthorCopyWith<$Res> {
+  factory $BookAuthorCopyWith(
+          BookAuthor value, $Res Function(BookAuthor) then) =
+      _$BookAuthorCopyWithImpl<$Res>;
   $Res call({int id, String? name, HasMany<Book>? books});
 }
 
 /// @nodoc
-class _$AuthorCopyWithImpl<$Res> implements $AuthorCopyWith<$Res> {
-  _$AuthorCopyWithImpl(this._value, this._then);
+class _$BookAuthorCopyWithImpl<$Res> implements $BookAuthorCopyWith<$Res> {
+  _$BookAuthorCopyWithImpl(this._value, this._then);
 
-  final Author _value;
+  final BookAuthor _value;
   // ignore: unused_field
-  final $Res Function(Author) _then;
+  final $Res Function(BookAuthor) _then;
 
   @override
   $Res call({
@@ -86,21 +88,23 @@ class _$AuthorCopyWithImpl<$Res> implements $AuthorCopyWith<$Res> {
 }
 
 /// @nodoc
-abstract class _$AuthorCopyWith<$Res> implements $AuthorCopyWith<$Res> {
-  factory _$AuthorCopyWith(_Author value, $Res Function(_Author) then) =
-      __$AuthorCopyWithImpl<$Res>;
+abstract class _$BookAuthorCopyWith<$Res> implements $BookAuthorCopyWith<$Res> {
+  factory _$BookAuthorCopyWith(
+          _BookAuthor value, $Res Function(_BookAuthor) then) =
+      __$BookAuthorCopyWithImpl<$Res>;
   @override
   $Res call({int id, String? name, HasMany<Book>? books});
 }
 
 /// @nodoc
-class __$AuthorCopyWithImpl<$Res> extends _$AuthorCopyWithImpl<$Res>
-    implements _$AuthorCopyWith<$Res> {
-  __$AuthorCopyWithImpl(_Author _value, $Res Function(_Author) _then)
-      : super(_value, (v) => _then(v as _Author));
+class __$BookAuthorCopyWithImpl<$Res> extends _$BookAuthorCopyWithImpl<$Res>
+    implements _$BookAuthorCopyWith<$Res> {
+  __$BookAuthorCopyWithImpl(
+      _BookAuthor _value, $Res Function(_BookAuthor) _then)
+      : super(_value, (v) => _then(v as _BookAuthor));
 
   @override
-  _Author get _value => super._value as _Author;
+  _BookAuthor get _value => super._value as _BookAuthor;
 
   @override
   $Res call({
@@ -108,7 +112,7 @@ class __$AuthorCopyWithImpl<$Res> extends _$AuthorCopyWithImpl<$Res>
     Object? name = freezed,
     Object? books = freezed,
   }) {
-    return _then(_Author(
+    return _then(_BookAuthor(
       id: id == freezed
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
@@ -127,12 +131,12 @@ class __$AuthorCopyWithImpl<$Res> extends _$AuthorCopyWithImpl<$Res>
 
 /// @nodoc
 @JsonSerializable()
-@With.fromString('DataModel<Author>')
-class _$_Author with DataModel<Author> implements _Author {
-  _$_Author({required this.id, this.name, this.books});
+@With.fromString('DataModel<BookAuthor>')
+class _$_BookAuthor with DataModel<BookAuthor> implements _BookAuthor {
+  _$_BookAuthor({required this.id, this.name, this.books});
 
-  factory _$_Author.fromJson(Map<String, dynamic> json) =>
-      _$_$_AuthorFromJson(json);
+  factory _$_BookAuthor.fromJson(Map<String, dynamic> json) =>
+      _$_$_BookAuthorFromJson(json);
 
   @override
   final int id;
@@ -143,13 +147,13 @@ class _$_Author with DataModel<Author> implements _Author {
 
   @override
   String toString() {
-    return 'Author(id: $id, name: $name, books: $books)';
+    return 'BookAuthor(id: $id, name: $name, books: $books)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
-        (other is _Author &&
+        (other is _BookAuthor &&
             (identical(other.id, id) ||
                 const DeepCollectionEquality().equals(other.id, id)) &&
             (identical(other.name, name) ||
@@ -167,20 +171,21 @@ class _$_Author with DataModel<Author> implements _Author {
 
   @JsonKey(ignore: true)
   @override
-  _$AuthorCopyWith<_Author> get copyWith =>
-      __$AuthorCopyWithImpl<_Author>(this, _$identity);
+  _$BookAuthorCopyWith<_BookAuthor> get copyWith =>
+      __$BookAuthorCopyWithImpl<_BookAuthor>(this, _$identity);
 
   @override
   Map<String, dynamic> toJson() {
-    return _$_$_AuthorToJson(this);
+    return _$_$_BookAuthorToJson(this);
   }
 }
 
-abstract class _Author implements Author, DataModel<Author> {
-  factory _Author({required int id, String? name, HasMany<Book>? books}) =
-      _$_Author;
+abstract class _BookAuthor implements BookAuthor, DataModel<BookAuthor> {
+  factory _BookAuthor({required int id, String? name, HasMany<Book>? books}) =
+      _$_BookAuthor;
 
-  factory _Author.fromJson(Map<String, dynamic> json) = _$_Author.fromJson;
+  factory _BookAuthor.fromJson(Map<String, dynamic> json) =
+      _$_BookAuthor.fromJson;
 
   @override
   int get id => throw _privateConstructorUsedError;
@@ -190,7 +195,8 @@ abstract class _Author implements Author, DataModel<Author> {
   HasMany<Book>? get books => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$AuthorCopyWith<_Author> get copyWith => throw _privateConstructorUsedError;
+  _$BookAuthorCopyWith<_BookAuthor> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 Book _$BookFromJson(Map<String, dynamic> json) {
@@ -201,11 +207,17 @@ Book _$BookFromJson(Map<String, dynamic> json) {
 class _$BookTearOff {
   const _$BookTearOff();
 
-  _Book call({required int id, String? title, BelongsTo<Author>? author}) {
+  _Book call(
+      {required int id,
+      String? title,
+      int numberOfSales = 0,
+      @JsonKey(name: 'original_author')
+          BelongsTo<BookAuthor>? originalAuthor}) {
     return _Book(
       id: id,
       title: title,
-      author: author,
+      numberOfSales: numberOfSales,
+      originalAuthor: originalAuthor,
     );
   }
 
@@ -221,7 +233,10 @@ const $Book = _$BookTearOff();
 mixin _$Book {
   int get id => throw _privateConstructorUsedError;
   String? get title => throw _privateConstructorUsedError;
-  BelongsTo<Author>? get author => throw _privateConstructorUsedError;
+  int get numberOfSales => throw _privateConstructorUsedError;
+  @JsonKey(name: 'original_author')
+  BelongsTo<BookAuthor>? get originalAuthor =>
+      throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
@@ -232,7 +247,11 @@ mixin _$Book {
 abstract class $BookCopyWith<$Res> {
   factory $BookCopyWith(Book value, $Res Function(Book) then) =
       _$BookCopyWithImpl<$Res>;
-  $Res call({int id, String? title, BelongsTo<Author>? author});
+  $Res call(
+      {int id,
+      String? title,
+      int numberOfSales,
+      @JsonKey(name: 'original_author') BelongsTo<BookAuthor>? originalAuthor});
 }
 
 /// @nodoc
@@ -247,7 +266,8 @@ class _$BookCopyWithImpl<$Res> implements $BookCopyWith<$Res> {
   $Res call({
     Object? id = freezed,
     Object? title = freezed,
-    Object? author = freezed,
+    Object? numberOfSales = freezed,
+    Object? originalAuthor = freezed,
   }) {
     return _then(_value.copyWith(
       id: id == freezed
@@ -258,10 +278,14 @@ class _$BookCopyWithImpl<$Res> implements $BookCopyWith<$Res> {
           ? _value.title
           : title // ignore: cast_nullable_to_non_nullable
               as String?,
-      author: author == freezed
-          ? _value.author
-          : author // ignore: cast_nullable_to_non_nullable
-              as BelongsTo<Author>?,
+      numberOfSales: numberOfSales == freezed
+          ? _value.numberOfSales
+          : numberOfSales // ignore: cast_nullable_to_non_nullable
+              as int,
+      originalAuthor: originalAuthor == freezed
+          ? _value.originalAuthor
+          : originalAuthor // ignore: cast_nullable_to_non_nullable
+              as BelongsTo<BookAuthor>?,
     ));
   }
 }
@@ -271,7 +295,11 @@ abstract class _$BookCopyWith<$Res> implements $BookCopyWith<$Res> {
   factory _$BookCopyWith(_Book value, $Res Function(_Book) then) =
       __$BookCopyWithImpl<$Res>;
   @override
-  $Res call({int id, String? title, BelongsTo<Author>? author});
+  $Res call(
+      {int id,
+      String? title,
+      int numberOfSales,
+      @JsonKey(name: 'original_author') BelongsTo<BookAuthor>? originalAuthor});
 }
 
 /// @nodoc
@@ -287,7 +315,8 @@ class __$BookCopyWithImpl<$Res> extends _$BookCopyWithImpl<$Res>
   $Res call({
     Object? id = freezed,
     Object? title = freezed,
-    Object? author = freezed,
+    Object? numberOfSales = freezed,
+    Object? originalAuthor = freezed,
   }) {
     return _then(_Book(
       id: id == freezed
@@ -298,19 +327,28 @@ class __$BookCopyWithImpl<$Res> extends _$BookCopyWithImpl<$Res>
           ? _value.title
           : title // ignore: cast_nullable_to_non_nullable
               as String?,
-      author: author == freezed
-          ? _value.author
-          : author // ignore: cast_nullable_to_non_nullable
-              as BelongsTo<Author>?,
+      numberOfSales: numberOfSales == freezed
+          ? _value.numberOfSales
+          : numberOfSales // ignore: cast_nullable_to_non_nullable
+              as int,
+      originalAuthor: originalAuthor == freezed
+          ? _value.originalAuthor
+          : originalAuthor // ignore: cast_nullable_to_non_nullable
+              as BelongsTo<BookAuthor>?,
     ));
   }
 }
 
 /// @nodoc
-@JsonSerializable()
+
 @With.fromString('DataModel<Book>')
+@JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class _$_Book with DataModel<Book> implements _Book {
-  _$_Book({required this.id, this.title, this.author});
+  _$_Book(
+      {required this.id,
+      this.title,
+      this.numberOfSales = 0,
+      @JsonKey(name: 'original_author') this.originalAuthor});
 
   factory _$_Book.fromJson(Map<String, dynamic> json) =>
       _$_$_BookFromJson(json);
@@ -319,12 +357,16 @@ class _$_Book with DataModel<Book> implements _Book {
   final int id;
   @override
   final String? title;
+  @JsonKey(defaultValue: 0)
   @override
-  final BelongsTo<Author>? author;
+  final int numberOfSales;
+  @override
+  @JsonKey(name: 'original_author')
+  final BelongsTo<BookAuthor>? originalAuthor;
 
   @override
   String toString() {
-    return 'Book(id: $id, title: $title, author: $author)';
+    return 'Book(id: $id, title: $title, numberOfSales: $numberOfSales, originalAuthor: $originalAuthor)';
   }
 
   @override
@@ -335,8 +377,12 @@ class _$_Book with DataModel<Book> implements _Book {
                 const DeepCollectionEquality().equals(other.id, id)) &&
             (identical(other.title, title) ||
                 const DeepCollectionEquality().equals(other.title, title)) &&
-            (identical(other.author, author) ||
-                const DeepCollectionEquality().equals(other.author, author)));
+            (identical(other.numberOfSales, numberOfSales) ||
+                const DeepCollectionEquality()
+                    .equals(other.numberOfSales, numberOfSales)) &&
+            (identical(other.originalAuthor, originalAuthor) ||
+                const DeepCollectionEquality()
+                    .equals(other.originalAuthor, originalAuthor)));
   }
 
   @override
@@ -344,7 +390,8 @@ class _$_Book with DataModel<Book> implements _Book {
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(id) ^
       const DeepCollectionEquality().hash(title) ^
-      const DeepCollectionEquality().hash(author);
+      const DeepCollectionEquality().hash(numberOfSales) ^
+      const DeepCollectionEquality().hash(originalAuthor);
 
   @JsonKey(ignore: true)
   @override
@@ -358,8 +405,12 @@ class _$_Book with DataModel<Book> implements _Book {
 }
 
 abstract class _Book implements Book, DataModel<Book> {
-  factory _Book({required int id, String? title, BelongsTo<Author>? author}) =
-      _$_Book;
+  factory _Book(
+      {required int id,
+      String? title,
+      int numberOfSales,
+      @JsonKey(name: 'original_author')
+          BelongsTo<BookAuthor>? originalAuthor}) = _$_Book;
 
   factory _Book.fromJson(Map<String, dynamic> json) = _$_Book.fromJson;
 
@@ -368,7 +419,11 @@ abstract class _Book implements Book, DataModel<Book> {
   @override
   String? get title => throw _privateConstructorUsedError;
   @override
-  BelongsTo<Author>? get author => throw _privateConstructorUsedError;
+  int get numberOfSales => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(name: 'original_author')
+  BelongsTo<BookAuthor>? get originalAuthor =>
+      throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
   _$BookCopyWith<_Book> get copyWith => throw _privateConstructorUsedError;