+++ test/_support/node.g.dart
@@ -123,9 +123,15 @@ AutoDisposeStateNotifierProvider<DataStateNotifier<List<Node>>,
     watchNodes(
         {bool? remote,
         Map<String, dynamic>? params,
-        Map<String, String>? headers}) {
-  return _watchNodes(
-      WatchArgs(remote: remote, params: params, headers: headers));
+        Map<String, String>? headers,
+        bool Function(Node)? filterLocal,
+        bool? syncLocal}) {
+  return _watchNodes(WatchArgs(
+      remote: remote,
+      params: params,
+      headers: headers,
+      filterLocal: filterLocal,
+      syncLocal: syncLocal));
 }
 
 extension NodeX on Node {
@@ -133,8 +139,10 @@ extension NodeX on Node {
   /// [save], [delete] and so on.
   ///
   /// Can be obtained via `context.read`, `ref.read`, `container.read`
-  Node init(Reader read) {
+  Node init(Reader read, {bool save = true}) {
     final repository = internalLocatorFn(nodesRepositoryProvider, read);
-    return repository.remoteAdapter.initializeModel(this, save: true);
+    final updatedModel =
+        repository.remoteAdapter.initializeModel(this, save: save);
+    return save ? updatedModel : this;
   }
 }