+++ test/_support/person.g.dart
@@ -97,9 +97,15 @@ AutoDisposeStateNotifierProvider<DataStateNotifier<List<Person>>,
     watchPeople(
         {bool? remote,
         Map<String, dynamic>? params,
-        Map<String, String>? headers}) {
-  return _watchPeople(
-      WatchArgs(remote: remote, params: params, headers: headers));
+        Map<String, String>? headers,
+        bool Function(Person)? filterLocal,
+        bool? syncLocal}) {
+  return _watchPeople(WatchArgs(
+      remote: remote,
+      params: params,
+      headers: headers,
+      filterLocal: filterLocal,
+      syncLocal: syncLocal));
 }
 
 extension PersonX on Person {
@@ -107,8 +113,10 @@ extension PersonX on Person {
   /// [save], [delete] and so on.
   ///
   /// Can be obtained via `context.read`, `ref.read`, `container.read`
-  Person init(Reader read) {
+  Person init(Reader read, {bool save = true}) {
     final repository = internalLocatorFn(peopleRepositoryProvider, read);
-    return repository.remoteAdapter.initializeModel(this, save: true);
+    final updatedModel =
+        repository.remoteAdapter.initializeModel(this, save: save);
+    return save ? updatedModel : this;
   }
 }